Reconnector redux (#40)
* Lots of progress to rework reconnector

There are a bunch of race conditions and restrictions that make no sense

* Fix controller unit tests and bug with tseventemitter

* Bunch of tiny fixes + dont use mutex implicitly when requiring config, wth

* Bump build

* Force ipv4

* Bump build

* Update joltron version, fix bug with awaiting on message sent vs message response

* bump build

* 2.7.0

* Update winston to 3.6.0

Smaller dependencies and pinned version of color.js

* 2.7.1

* Bump joltron version to 2.2.9-beta.2

* 2.7.2

* Make sure not to spread on undefined/nulls

* Bump build

* 2.7.3

+++ src/shortcut.ts
@@ -23,6 +23,10 @@ export abstract class Shortcut {
 	}
 
 	private static async createLinux(program: string, icon: string) {
+		if (!xdgBasedir.data) {
+			throw new Error('Could not resolve desktop shortcut dir using XDG');
+		}
+
 		let desktopFile = path.join(
 			xdgBasedir.data || '',
 			'applications',
@@ -47,6 +51,10 @@ Name[en_US]=Game Jolt Client`;
 	}
 
 	private static removeLinux() {
+		if (!xdgBasedir.data) {
+			throw new Error('Could not resolve desktop shortcut dir using XDG');
+		}
+
 		let desktopFile = path.join(xdgBasedir.data, 'applications', 'game-jolt-client.desktop');
 		let oldDesktopFile = path.join(xdgBasedir.data, 'applications', 'Game Jolt Client.desktop');
 		return Promise.all([fs.unlink(desktopFile), fs.unlink(oldDesktopFile)])
